First commit of basic Smart Contract
This is the current base for the Smart Contract and how it is used on the testnet. Therfore the check if a wallet is a current custodian is commented out.
This SC allows current custodians of a planet to add or update a constitution, add, update and delete documents as well as start events on the planet.